[CI] Fix static-check, clang-format, Tizen/Windows-ARM narrowing and temp-path failures
Static checks: add @author, @bug, @see tags to vjepa_layernorm_layer.cpp and add @brief doxygen comment to VjepaLayerNormLayer class in vjepa_layernorm_layer.h so doxygen-tag.sh advanced checks pass.

Tizen GBS armv7l compile error: safetensors_header.cpp line 260 initialised Cursor::end (size_t) from header_size (uint64_t). On 32-bit ARM, size_t is uint32_t causing -Werror=narrowing. The narrowing is safe because line 250 already validates 8+header_size<=file_size<=SIZE_MAX, so cast with static_cast<size_t>.

Windows ARM tests: unittest_nntrainer_safetensors_header and unittest_nntrainer_weight_source used hardcoded /tmp/ paths that do not exist on Windows. Replace with std::filesystem::temp_directory_path() which resolves to the platform temp directory on both Linux and Windows.

clang-format-14: apply to the exact changed-line ranges flagged by the cpp-linter CI check (23 files, lines-changed-only mode matching cpp_linter.yml). Files already compliant were skipped.

+++ nntrainer/utils/safetensors_header.cpp
@@ -257,7 +257,8 @@ SafetensorsHeader parseSafetensorsHeader(const void *data, size_t size) {
   hdr.header_size = header_size;
   hdr.data_block_offset = 8 + header_size;
 
-  Cursor c{reinterpret_cast<const char *>(p + 8), 0, header_size};
+  Cursor c{reinterpret_cast<const char *>(p + 8), 0,
+           static_cast<size_t>(header_size)};
   expect(c, '{');
   while (true) {
     skipWs(c);
